# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "REST API for Topcoder Submissions.\n\n### Pagination\n  Requests that\
    \ return multiple items will be paginated to 20 items by default. You can specify\n\
    \  further pages with the `page` parameter. You can also set a custom page\n \
    \ size up to 100 with the `perPage` parameter.\n\n  Pagination response data is\
    \ included in http headers. By Default, the response header contains links with\
    \ `next`, `last`, `first`, `prev` resource links.\n"
  version: "5.0.0"
  title: "Topcoder Submissions REST API"
  termsOfService: "http://topcoder.com/terms/"
  contact:
    url: "http://www.topcoder.com/support"
    email: "support@topcoder.com"
host: "api.topcoder.com"
basePath: "/v5"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
security:
- Bearer: []
paths:
...
  /reviews:
    get:
      tags:
      - "Reviews"
      summary: "Get all reviews."
      description: "Get all reviews. Link headers are sent back and they have rel\
        \ set to prev, next, first, last and contain the relevant URL.\n\nResulted\
        \ collection of reviews can be filtered using filter parameters `score`, `typeId`,\
        \ `reviewerId`, `scoreCardId` (all filter parameters are optional and combined\
        \ by the logical operation `AND`).\n\n**Authorization:** Review is accessible\
        \ by roles `admin` and `copilot`.\n"
      operationId: "getReviews"
      parameters:
      - name: "page"
        in: "query"
        description: "The page number."
        required: false
        type: "integer"
        default: 1
      - name: "perPage"
        in: "query"
        description: "The number of items to list per page."
        required: false
        type: "integer"
        default: 20
        maximum: 100
      - name: "score"
        in: "query"
        description: "The score filter for reviews."
        required: false
        type: "number"
        format: "double"
      - name: "typeId"
        in: "query"
        description: "The type id filter for reviews."
        required: false
        type: "string"
      - name: "reviewerId"
        in: "query"
        description: "The reviewer id filter for reviews."
        required: false
        type: "string"
      - name: "scoreCardId"
        in: "query"
        description: "The score card id filter for reviews."
        required: false
        type: "string"
      - name: "If-None-Match"
        in: "header"
        description: "Optional. If used in the request, the If-None-Match request\
          \ header should utilize the ETag as previously supplied in the response.\n"
        required: false
        type: "string"
      - name: "If-Modified-Since"
        in: "header"
        description: "Optional. In the subsequent request, the If-Modified-Since request\
          \ header can be populated with the previously retrieved `Last-Modified`.\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK - the request was successful, the resource are returned."
          schema:
            type: "array"
            example:
            - id: "a75529df08beff833f4c6282"
              score: 95.5
              typeId: "b0b2a8220fcd169f684093fb"
              reviewerId: "fcadaf16dae64f5fb43bc7ee"
              scoreCardId: "0bf26a9e69e564a6c2988d12"
              submissionId: "255456"
              created: "2018-05-20T07:00:30.123Z"
              updated: "2018-06-01T07:36:28.178Z"
              createdBy: "admin"
              updatedBy: "admin"
            - id: "e84031c74a9566fe5ab848eb"
              score: 77.8
              typeId: "8bc66db12b81f038a19ecd00"
              reviewerId: "973fd66d229a7cfdbe9d9981"
              scoreCardId: "ffd96446bda80799bb2d34ed"
              submissionId: "255457"
              created: "2018-05-20T07:00:30.123Z"
              updated: "2018-06-01T07:36:28.178Z"
              createdBy: "admin"
              updatedBy: "admin"
            items:
              $ref: "#/definitions/Review"
          headers:
            X-Next-Page:
              type: "integer"
              description: "The index of the next page"
            X-Page:
              type: "integer"
              description: "The index of the current page (starting at 1)"
            X-Per-Page:
              type: "integer"
              description: "The number of items to list per page"
            X-Prev-Page:
              type: "integer"
              description: "The index of the previous page"
            X-Total:
              type: "integer"
              description: "The total number of items"
            X-Total-Pages:
              type: "integer"
              description: "The total number of pages"
            ETag:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                Common methods of ETag generation include using a collision-resistant\
                \ hash function of the resource's content, a hash of the last modification\
                \ timestamp, or even just a revision number.\n\nIn the next request,\
                \ the If-None-Match request header should utilize the ETag as previously\
                \ supplied in the response.\n"
            Last-Modified:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                The Last-Modified response HTTP header contains the date and time\
                \ at which the origin server believes the resource was last modified.\
                \ It is used as a validator to determine if a resource received or\
                \ stored is the same. Less accurate than an ETag header, it is a fallback\
                \ mechanism.\n\nNormally a database or otherwise maintained `Last-Modified\
                \ date` is maintained for the resource, which is supplied in the Last-Modified\
                \ response header.\n\nIn the subsequent request, the If-Modified-Since\
                \ request header can be populated with the previously retrieved `Last-Modified`.\n"
            Link:
              type: "string"
              description: "Pagination link header."
        304:
          $ref: "#/definitions/NotModified"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "Reviews"
    head:
      tags:
      - "Reviews"
      summary: "Get only response status and headers information but no response body\
        \ for the endpoint.\n"
      description: "Get response status and headers information for the endpoint.\
        \ Link headers are sent back and they have rel set to prev, next, first, last\
        \ and contain the relevant URL.\n\nRequested reviews can be filtered using\
        \ filter parameters `score`, `typeId`, `reviewerId`, `scoreCardId` (all filter\
        \ parameters are optional and combined by the logical operation `AND`).\n\n\
        **Authorization:** Review is accessible by roles `admin` and `copilot`.\n"
      operationId: "headReviews"
      parameters:
      - name: "page"
        in: "query"
        description: "The page number."
        required: false
        type: "integer"
        default: 1
      - name: "perPage"
        in: "query"
        description: "The number of items to list per page."
        required: false
        type: "integer"
        default: 20
        maximum: 100
      - name: "score"
        in: "query"
        description: "The score filter for reviews."
        required: false
        type: "number"
        format: "double"
      - name: "typeId"
        in: "query"
        description: "The type id filter for reviews."
        required: false
        type: "string"
      - name: "reviewerId"
        in: "query"
        description: "The reviewer id filter for reviews."
        required: false
        type: "string"
      - name: "scoreCardId"
        in: "query"
        description: "The score card id filter for reviews."
        required: false
        type: "string"
      - name: "If-None-Match"
        in: "header"
        description: "Optional. If used in the request, the If-None-Match request\
          \ header should utilize the ETag as previously supplied in the response.\n"
        required: false
        type: "string"
      - name: "If-Modified-Since"
        in: "header"
        description: "Optional. In the subsequent request, the If-Modified-Since request\
          \ header can be populated with the previously retrieved `Last-Modified`.\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK - The request was successful."
          headers:
            X-Next-Page:
              type: "integer"
              description: "The index of the next page"
            X-Page:
              type: "integer"
              description: "The index of the current page (starting at 1)"
            X-Per-Page:
              type: "integer"
              description: "The number of items to list per page"
            X-Prev-Page:
              type: "integer"
              description: "The index of the previous page"
            X-Total:
              type: "integer"
              description: "The total number of items"
            X-Total-Pages:
              type: "integer"
              description: "The total number of pages"
            ETag:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                Common methods of ETag generation include using a collision-resistant\
                \ hash function of the resource's content, a hash of the last modification\
                \ timestamp, or even just a revision number.\n\nIn the next request,\
                \ the If-None-Match request header should utilize the ETag as previously\
                \ supplied in the response.\n"
            Last-Modified:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                The Last-Modified response HTTP header contains the date and time\
                \ at which the origin server believes the resource was last modified.\
                \ It is used as a validator to determine if a resource received or\
                \ stored is the same. Less accurate than an ETag header, it is a fallback\
                \ mechanism.\n\nNormally a database or otherwise maintained `Last-Modified\
                \ date` is maintained for the resource, which is supplied in the Last-Modified\
                \ response header.\n\nIn the subsequent request, the If-Modified-Since\
                \ request header can be populated with the previously retrieved `Last-Modified`.\n"
            Link:
              type: "string"
              description: "Pagination link header."
        304:
          $ref: "#/definitions/NotModifiedWithoutBody"
        400:
          $ref: "#/definitions/BadRequestWithoutBody"
        401:
          $ref: "#/definitions/UnauthorizedWithoutBody"
        403:
          $ref: "#/definitions/ForbiddenWithoutBody"
        500:
          $ref: "#/definitions/ServerErrorWithoutBody"
      x-swagger-router-controller: "Reviews"
    post:
      tags:
      - "Reviews"
      summary: "Create a review."
      description: "Create a new review.\n\n**Authorization:** Review creation is\
        \ accessible by roles `admin` and `copilot`.\n"
      operationId: "createReview"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Review"
      responses:
        201:
          description: "Created - The request was successful and the resource is returned."
          schema:
            $ref: "#/definitions/Review"
          headers:
            Location:
              type: "string"
              description: "The Location response header indicates the URL to redirect\
                \ a page to or created resource. It only provides a meaning when served\
                \ with a 3xx (redirection) or 201 (created) status response."
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        409:
          $ref: "#/definitions/Conflict"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "Reviews"
  /reviews/{reviewId}:
    get:
      tags:
      - "Reviews"
      summary: "Get the review."
      description: "Get the review by id.\n\n**Authorization:** Review is accessible\
        \ by roles `admin` and `copilot`.\n"
      operationId: "getReview"
      parameters:
      - name: "reviewId"
        in: "path"
        description: "review id"
        required: true
        type: "string"
      - name: "If-None-Match"
        in: "header"
        description: "Optional. If used in the request, the If-None-Match request\
          \ header should utilize the ETag as previously supplied in the response.\n"
        required: false
        type: "string"
      - name: "If-Modified-Since"
        in: "header"
        description: "Optional. In the subsequent request, the If-Modified-Since request\
          \ header can be populated with the previously retrieved `Last-Modified`.\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK - the request was successful and the review is returned."
          schema:
            $ref: "#/definitions/Review"
          headers:
            ETag:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                Common methods of ETag generation include using a collision-resistant\
                \ hash function of the resource's content, a hash of the last modification\
                \ timestamp, or even just a revision number.\n\nIn the next request,\
                \ the If-None-Match request header should utilize the ETag as previously\
                \ supplied in the response.\n"
            Last-Modified:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                The Last-Modified response HTTP header contains the date and time\
                \ at which the origin server believes the resource was last modified.\
                \ It is used as a validator to determine if a resource received or\
                \ stored is the same. Less accurate than an ETag header, it is a fallback\
                \ mechanism.\n\nNormally a database or otherwise maintained `Last-Modified\
                \ date` is maintained for the resource, which is supplied in the Last-Modified\
                \ response header.\n\nIn the subsequent request, the If-Modified-Since\
                \ request header can be populated with the previously retrieved `Last-Modified`.\n"
        304:
          $ref: "#/definitions/NotModified"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        404:
          $ref: "#/definitions/NotFound"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "Reviews"
    head:
      tags:
      - "Reviews"
      summary: "Get only response status and headers information but no response body\
        \ for the endpoint.\n"
      description: "Get response status and headers information for the endpoint.\
        \ It does not contain response body.\n\n**Authorization:** Review is accessible\
        \ by roles `admin` and `copilot`.\n"
      operationId: "headReview"
      parameters:
      - name: "reviewId"
        in: "path"
        description: "review id"
        required: true
        type: "string"
      - name: "If-None-Match"
        in: "header"
        description: "Optional. If used in the request, the If-None-Match request\
          \ header should utilize the ETag as previously supplied in the response.\n"
        required: false
        type: "string"
      - name: "If-Modified-Since"
        in: "header"
        description: "Optional. In the subsequent request, the If-Modified-Since request\
          \ header can be populated with the previously retrieved `Last-Modified`.\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK - The request was successful."
          headers:
            ETag:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                Common methods of ETag generation include using a collision-resistant\
                \ hash function of the resource's content, a hash of the last modification\
                \ timestamp, or even just a revision number.\n\nIn the next request,\
                \ the If-None-Match request header should utilize the ETag as previously\
                \ supplied in the response.\n"
            Last-Modified:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                The Last-Modified response HTTP header contains the date and time\
                \ at which the origin server believes the resource was last modified.\
                \ It is used as a validator to determine if a resource received or\
                \ stored is the same. Less accurate than an ETag header, it is a fallback\
                \ mechanism.\n\nNormally a database or otherwise maintained `Last-Modified\
                \ date` is maintained for the resource, which is supplied in the Last-Modified\
                \ response header.\n\nIn the subsequent request, the If-Modified-Since\
                \ request header can be populated with the previously retrieved `Last-Modified`.\n"
        304:
          $ref: "#/definitions/NotModifiedWithoutBody"
        400:
          $ref: "#/definitions/BadRequestWithoutBody"
        401:
          $ref: "#/definitions/UnauthorizedWithoutBody"
        403:
          $ref: "#/definitions/ForbiddenWithoutBody"
        500:
          $ref: "#/definitions/ServerErrorWithoutBody"
      x-swagger-router-controller: "Reviews"
    put:
      tags:
      - "Reviews"
      summary: "Update the review."
      description: "Update the review by id.\n\n**Authorization:** Review update is\
        \ accessible only by `admin` role.\n"
      operationId: "updateReview"
      parameters:
      - name: "reviewId"
        in: "path"
        description: "review id"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/UpdatableReview"
      responses:
        200:
          description: "OK - The request was successful and the updated resource is\
            \ returned."
          schema:
            $ref: "#/definitions/Review"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        404:
          $ref: "#/definitions/NotFound"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "Reviews"
    delete:
      tags:
      - "Reviews"
      summary: "Delete the review."
      description: "Delete the review.\n\n**Authorization:** Review deletion is accessible\
        \ only by `admin` role.\n"
      operationId: "deleteReview"
      parameters:
      - name: "reviewId"
        in: "path"
        description: "review id"
        required: true
        type: "string"
      responses:
        204:
          $ref: "#/definitions/NoContent"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        404:
          $ref: "#/definitions/NotFound"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "Reviews"
    patch:
      tags:
      - "Reviews"
      summary: "Partially update the review."
      description: "Allows to partially modify the review with the provided request\
        \ body properties.\n\n**Authorization:** Partially modify of review is accessible\
        \ only by `admin` role.\n"
      operationId: "partiallyUpdateReview"
      parameters:
      - name: "reviewId"
        in: "path"
        description: "review id"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/PartiallyReview"
      responses:
        200:
          description: "OK - The request was successful and the updated resource is\
            \ returned."
          schema:
            $ref: "#/definitions/Review"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        404:
          $ref: "#/definitions/NotFound"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "Reviews"
...
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  NoContent:
    description: "No Content - The server has successfully fulfilled the request and\
      \ that there is no additional content to send in the response payload body.\n"
  BadRequestWithoutBody:
    description: "Bad Request - Request validation error, unable to process the request\
      \ due to an error (e.g., malformed request syntax, size too large, invalid request\
      \ message framing, or deceptive request routing).\n"
  UnauthorizedWithoutBody:
    description: "Unauthorized - Occurs when the JWT token provided is either expired\
      \ or invalid.\n"
  ForbiddenWithoutBody:
    description: "Forbidden - The request is not allowed.\n"
  NotModifiedWithoutBody:
    description: "Not Modified - Indicates that the resource has not been modified\
      \ since the last request.\n"
  ServerErrorWithoutBody:
    description: "Server Error - While handling the request something went wrong server-side.\n"
...
  Review:
    allOf:
    - type: "object"
      properties:
        id:
          type: "string"
          example: "a75529df08beff833f4c6282"
          description: "The review id."
          readOnly: true
    - $ref: "#/definitions/PartiallyReview"
    - $ref: "#/definitions/Auditable"
    description: "The review entity."
  PartiallyReview:
    type: "object"
    properties:
      score:
        type: "number"
        format: "double"
        example: 95.5
        description: "The review score."
      typeId:
        type: "string"
        example: "b0b2a8220fcd169f684093fb"
        description: "The review type id."
      reviewerId:
        type: "string"
        example: "fcadaf16dae64f5fb43bc7ee"
        description: "The review reviewer id."
      scoreCardId:
        type: "string"
        example: "0bf26a9e69e564a6c2988d12"
        description: "The review score card id."
      submissionId:
        type: "string"
        example: "255456"
        description: "The submission id."
    description: "The review entity fields that can be partially updated."
    example:
      score: 95.5
      reviewerId: "fcadaf16dae64f5fb43bc7ee"
      submissionId: "255456"
      scoreCardId: "0bf26a9e69e564a6c2988d12"
      typeId: "b0b2a8220fcd169f684093fb"
  UpdatableReview:
    allOf:
    - $ref: "#/definitions/PartiallyReview"
    description: "The review entity fields that updates whole entity."
...
  Auditable:
    type: "object"
    properties:
      created:
        type: "string"
        format: "date-time"
        example: "2018-05-20T07:00:30.123Z"
        description: "The entity creation date."
        readOnly: true
      updated:
        type: "string"
        format: "date-time"
        example: "2018-06-01T07:36:28.178Z"
        description: "The entity update date."
        readOnly: true
      createdBy:
        type: "string"
        example: "admin"
        description: "The user who created the entity."
        readOnly: true
      updatedBy:
        type: "string"
        example: "admin"
        description: "The user who updated the entity."
        readOnly: true
    description: "The entity used as mark to add audit fields."
  Unauthorized:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Unable to authenticate the user."
        description: "The unauthorized error message."
    description: "The unauthorized error entity."
  NotFound:
    type: "object"
    properties:
      message:
        type: "string"
        example: "A resource with the name could not be found."
        description: "The not found error message."
    description: "The not found error entity."
  NotModified:
    type: "object"
    properties:
      message:
        type: "string"
        example: "There is no need to transfer a representation of the target resource\
          \ because the request indicates that the client, which made the request\
          \ conditional, already has a valid representation."
        description: "The not modified error message."
    description: "The not modified entity."
  ServerError:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Something went wrong while processing your request. We’re sorry\
          \ for the trouble. We’ve been notified of the error and will correct it\
          \ as soon as possible. Please try your request again in a moment."
        description: "The server error message."
    description: "The server error entity."
  BadRequest:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Invalid input."
        description: "The bad request error message."
    description: "The bad request error entity."
  Forbidden:
    type: "object"
    properties:
      message:
        type: "string"
        example: "You are not allowed to access the request."
        description: "The forbidden error message."
    description: "The permission error entity."
  Conflict:
    type: "object"
    required:
    - "message"
    properties:
      message:
        type: "string"
        example: "Creating a resource with a name already exists."
        description: "The conflict error message."
    description: "The conflict error entity."
...